refactor(sema): move compiler outputs
