+++ crates/sema/Cargo.toml
@@ -35,6 +35,7 @@ num-bigint.workspace = true
 num-traits.workspace = true
 once_map.workspace = true
 rayon.workspace = true
+serde = { workspace = true, features = ["derive"] }
 thread_local.workspace = true
 tracing.workspace = true
